fix: 오타 수정

+++ server/src/main/java/com/seniclass/server/domain/student/repository/AssignmentSubmissionRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssignmentSubmissionRepository
-        extends JpaRepository<AssignmentSubmission, Integer> {}
+public interface AssignmentSubmissionRepository extends JpaRepository<AssignmentSubmission, Long> {}
