+++ server/src/main/java/com/seniclass/server/domain/student/repository/VideoProgressRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VideoProgressRepository extends JpaRepository<VideoProgress, Integer> {}
+public interface VideoProgressRepository extends JpaRepository<VideoProgress, Long> {}
